Move models from extensions into a models folder
